+++ NEWS.md
@@ -19,6 +19,11 @@ title: "NEWS for the emmeans package"
   * Fixed a scoping issue in `contrast.emmGrid` to make a custom `.emmc` function
     easier to find. This bug prevented some examples from being rendered correctly
     in all contexts.
+  * Restored the `joint_tests()` code that was omitted in 1.11.0 because apparently
+    it was right the first time. (#528)
+  * Added an argument `npts` to `make.meanint()` and `make.symmint()` to facilitate
+    generating an interval with more than two points. 
+  * Fix to `test()` for situations with non-estimability and infinite df (#528)
   
 ## emmeans 1.11.0
   * Added a `linfct()` generic and default method that returns `object@linfct`